Plxsep 1666 multiple tags cause versioning issues (#7)
* updated manual usage section of the readme.md to capture all of the required options

* Refine git version discovery and tag handling

Ancestor Tag Search: If the current commit lacks a valid SemVer tag, the tool now intelligently searches the git history for the most recent ancestor commit that has one. This provides more consistent versioning on development branches that are ahead of a tag.

Strict Ambiguity Checks: To prevent non-deterministic builds, an error is now raised if a commit (whether it's the current HEAD or an ancestor) is associated with multiple valid SemVer tags. This enforces clear and unambiguous versioning.

* address ruff s603 warnings

* address ruff warrning about compute_version being too complex

* delete version file if plxsversion reports an error

* address ruff warnings by delaring certian methods that don't realy on state static

* update tool version

* update readme to correct typos

---------

Co-authored-by: Uren Patel <uren.patel@plexus.com>